+++ src/components/controls/tabs/main/BackgroundTab.tsx
@@ -382,6 +382,12 @@ export default function BackgroundTab({
 					onClearSpectrumOverride={() =>
 						store.setImageSpectrumOverride(null)
 					}
+					onCaptureSecondSpectrumOverride={
+						store.captureImageSecondSpectrumOverride
+					}
+					onClearSecondSpectrumOverride={() =>
+						store.setImageSecondSpectrumOverride(null)
+					}
 					onCaptureParticlesOverride={
 						store.captureImageParticlesOverride
 					}